# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jarvis Voice AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Processing mode
 */
export type DictationInputMode =
  (typeof DictationInputMode)[keyof typeof DictationInputMode];

export const DictationInputMode = {
  clean: "clean",
  rephrase: "rephrase",
  bullets: "bullets",
  assistant: "assistant",
} as const;

export interface DictationInput {
  /** Raw transcribed text from speech recognition */
  rawText: string;
  /** Processing mode */
  mode: DictationInputMode;
}

export interface DictationResult {
  processedText: string;
  mode: string;
}

export interface DictationEntry {
  id: number;
  rawText: string;
  processedText: string;
  mode: string;
  createdAt: string;
}

export interface DictationEntryInput {
  rawText: string;
  processedText: string;
  mode: string;
}

export interface DictationStats {
  totalDictations: number;
  totalWords: number;
  todayDictations: number;
  mostUsedMode: string;
}