# lib/api-zod/src/generated/types/dictationEntryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jarvis Voice AI API
 * OpenAPI spec version: 0.1.0
 */

export interface DictationEntryInput {
  rawText: string;
  processedText: string;
  mode: string;
}