# lib/api-zod/src/generated/types/dictationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jarvis Voice AI API
 * OpenAPI spec version: 0.1.0
 */
import type { DictationInputMode } from "./dictationInputMode";

export interface DictationInput {
  /** Raw transcribed text from speech recognition */
  rawText: string;
  /** Processing mode */
  mode: DictationInputMode;
}