# lib/api-zod/src/generated/types/dictationStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jarvis Voice AI API
 * OpenAPI spec version: 0.1.0
 */

export interface DictationStats {
  totalDictations: number;
  totalWords: number;
  todayDictations: number;
  mostUsedMode: string;
}